+++ src/05-Data_Collection.cuh
@@ -6,10 +6,10 @@
  */
 typedef struct SmallDataCollection {
     // small = 16 bit
-    unsigned float mulp[65536];
-    unsigned float warp[65536];
-    unsigned float lane[65536];
-    unsigned float time[65536];
+    float mulp[65536];
+    float warp[65536];
+    float lane[65536];
+    float time[65536];
 } SmallDataCollection;
 
 
@@ -18,10 +18,10 @@ typedef struct SmallDataCollection {
  */
 typedef struct MediumDataCollection {
     // medium = 24 bit
-    unsigned float mulp[16777216];
-    unsigned float warp[16777216];
-    unsigned float lane[16777216];
-    unsigned float time[16777216];
+    float mulp[16777216];
+    float warp[16777216];
+    float lane[16777216];
+    float time[16777216];
 } MediumDataCollection;
 
 
@@ -30,10 +30,10 @@ typedef struct MediumDataCollection {
  */
 typedef struct LargeDataCollection {
     // large = 32 bit
-    unsigned float mulp[4294967296];
-    unsigned float warp[4294967296];
-    unsigned float lane[4294967296];
-    unsigned float time[4294967296];
+    float mulp[4294967296];
+    float warp[4294967296];
+    float lane[4294967296];
+    float time[4294967296];
 } LargeDataCollection;
 
 #endif //GCPVE_04_DATA_COLLECTION_CUH
